+++ HappyPaws.API/Controllers/UsersController.cs
@@ -1,7 +1,10 @@
-﻿using HappyPaws.API.Contracts.DTOs.UserDTOs;
+﻿using HappyPaws.API.Auth.Policies;
+using HappyPaws.API.Contracts.DTOs.UserDTOs;
 using HappyPaws.Application.Interfaces;
 using HappyPaws.Core.Exceptions.Common;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Data;
 
 namespace HappyPaws.API.Controllers
 {
@@ -12,13 +15,16 @@ namespace HappyPaws.API.Controllers
     public class UsersController : ControllerBase
     {
         private readonly IUserService _usersService;
+        private readonly IAuthorizationService _authorizationService;
 
-        public UsersController(IUserService usersSerevice) 
+        public UsersController(IUserService usersSerevice, IAuthorizationService authorizationService) 
         {
             _usersService = usersSerevice;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
+        [Authorize(Roles = "Admin")]
         [ProducesResponseType(typeof(IEnumerable<UserDTO>), (StatusCodes.Status200OK))]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> GetAsync()
@@ -31,11 +37,19 @@ namespace HappyPaws.API.Controllers
         }
 
         [HttpGet("{id}")]
+        [Authorize]
         [ProducesResponseType(typeof(UserDTO), StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<IActionResult> GetByIdAsync(Guid id)
         {
             var user = await _usersService.GetAsync(id);
+
+            var authResult = await _authorizationService.AuthorizeAsync(User, user, PolicyNames.SameUser);
+
+            if (!authResult.Succeeded)
+            {
+                return Forbid();
+            }
 
             if (user == null) throw new ResourceNotFoundException();
 
@@ -43,6 +57,7 @@ namespace HappyPaws.API.Controllers
         }
 
         [HttpPut("{id}")]
+        [Authorize]
         [ProducesResponseType(typeof(UserDTO), StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
@@ -52,12 +67,20 @@ namespace HappyPaws.API.Controllers
 
             if (user == null) throw new ResourceNotFoundException();
 
+            var authResult = await _authorizationService.AuthorizeAsync(User, user, PolicyNames.SameUser);
+
+            if (!authResult.Succeeded)
+            {
+                return Forbid();
+            }
+
             var updated = await _usersService.UpdateAsync(id, UpdateUserDTO.ToDomain(userDTO));
 
             return Ok(UserDTO.FromDomain(updated));
         }
 
         [HttpDelete]
+        [Authorize]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
@@ -66,6 +89,13 @@ namespace HappyPaws.API.Controllers
             var user = _usersService.GetAsync(id);
 
             if (user == null) throw new ResourceNotFoundException();
+
+            var authResult = await _authorizationService.AuthorizeAsync(User, user, PolicyNames.SameUser);
+
+            if (!authResult.Succeeded)
+            {
+                return Forbid();
+            }
 
             await _usersService.DeleteAsync(id);
